Obtain source vocabularies with a given licensing restriction level.

Various UMLS source vocabularies have licensing restrictions, ranging from none (0) to level 4. See UMLS for details.

source_level(level, us = TRUE)

Arguments

level
The maximum restriction level.

us
Whether or not you are looking at US or non-US restriction levels.

Value

A list of source vocabularies that have restriction levels less than the entered level.

Examples

# Get all restriction-less US source vocabularies
source_level(0)

#>   [1] "AIR"           "AOD"           "AOT"           "ATC"          
#>   [5] "CCS"           "CHV"           "COSTAR"        "CSP"          
#>   [9] "CST"           "COSTART"       "CVX"           "DXP"          
#>  [13] "FMA"           "GO"            "HCPCS"         "HGNC"         
#>  [17] "HL7V2.5"       "HL7V3.0"       "HPO"           "ICD10PCS"     
#>  [21] "ICD9CM"        "ICPC"          "ICPCBAQ"       "ICPCDAN"      
#>  [25] "ICPCDUT"       "ICPCFIN"       "ICPCFRE"       "ICPCGER"      
#>  [29] "ICPCHEB"       "ICPCHUN"       "ICPCITA"       "ICPCNOR"      
#>  [33] "ICPCPOR"       "ICPCSPA"       "ICPCSWE"       "LCH"          
#>  [37] "LCH_NW"        "LNC"           "LNC-DE-CH"     "LNC-DE-DE"    
#>  [41] "LNC-EL-GR"     "LNC-ES-AR"     "LNC-ES-CH"     "LNC-ES-ES"    
#>  [45] "LNC-ET-EE"     "LNC-FR-BE"     "LNC-FR-CA"     "LNC-FR-CH"    
#>  [49] "LNC-FR-FR"     "LNC-IT-CH"     "LNC-IT-IT"     "LNC-KO-KR"    
#>  [53] "LNC-NL-NL"     "LNC-PT-BR"     "LNC-RU-RU"     "LNC-TR-TR"    
#>  [57] "LNC-ZH-CN"     "MCM"           "MEDLINEPLUS"   "MSH"          
#>  [61] "MTH"           "MTHCMSFRF"     "MTHHH"         "MTHICD9"      
#>  [65] "MTHMST"        "MTHMSTFRE"     "MTHMSTITA"     "MTHSPL"       
#>  [69] "MVX"           "NCBI"          "NCI"           "NCI_BioC"     
#>  [73] "NCI_BRIDG"     "NCI_CareLex"   "NCI_CDC"       "NCI_CDISC"    
#>  [77] "NCI_CRCH"      "NCI_CTCAE"     "NCI_CTEP-SDC"  "NCI_DCP"      
#>  [81] "NCI_DICOM"     "NCI_DTP"       "NCI_FDA"       "NCI_ICH"      
#>  [85] "NCI_JAX"       "NCI_KEGG"      "NCI_NCI-GLOSS" "NCI_NCI-HL7"  
#>  [89] "NCI_NCPDP"     "NCI_NICHD"     "NCI_PID"       "NCI_RENI"     
#>  [93] "NCI_UCUM"      "NCI_Zfin"      "NCISEER"       "NDFRT"        
#>  [97] "NDFRT_FDASPL"  "NDFRT_FMTSME"  "OMIM"          "PDQ"          
#> [101] "QMR"           "RAM"           "RXNORM"        "SOP"          
#> [105] "SPN"           "SRC"           "TKMT"          "USPMG"        
#> [109] "UWDA"          "VANDF"